# backend/openapi/app.yaml
...
"openapi": "3.1.0"
"info":
  "title": "Vigía — interfaz de la plataforma"
  "version": "0.1.0"
"paths":
...
  "/ledger/records":
    "get":
      "tags":
        - "expediente"
      "summary": "Registros del expediente con alcance, paginados por clave (lectura auditada)"
      "operationId": "list_records_ledger_records_get"
      "parameters":
        - "name": "record_type"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "array"
                "items":
                  "type": "string"
                  "pattern": "^[a-z][a-z0-9_]{0,63}$"
                "maxItems": 32
              - "type": "null"
            "title": "Record Type"
        - "name": "plant_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Plant Id"
        - "name": "zone_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Zone Id"
        - "name": "node_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Node Id"
        - "name": "received_from"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "Received From"
        - "name": "received_before"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "Received Before"
        - "name": "page_size"
          "in": "query"
          "required": false
          "schema":
            "type": "integer"
            "maximum": 200
            "minimum": 1
            "default": 50
            "title": "Page Size"
        - "name": "after_received_at"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "After Received At"
        - "name": "after_record_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "After Record Id"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/LedgerRecordPage"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/ledger/records/{record_id}":
    "get":
      "tags":
        - "expediente"
      "summary": "Un registro del expediente (lectura auditada); fuera de alcance, not_found"
      "operationId": "get_record_ledger_records__record_id__get"
      "parameters":
        - "name": "record_id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
            "title": "Record Id"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/LedgerRecordOut"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/audit/entries":
    "get":
      "tags":
        - "expediente"
      "summary": "Entradas de la cadena de auditoría con alcance (lectura auditada)"
      "operationId": "list_audit_entries_audit_entries_get"
      "parameters":
        - "name": "operation"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "array"
                "items":
                  "type": "string"
                  "pattern": "^[a-z][a-z0-9_]{0,63}$"
                "maxItems": 32
              - "type": "null"
            "title": "Operation"
        - "name": "actor_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Actor Id"
        - "name": "plant_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Plant Id"
        - "name": "zone_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "Zone Id"
        - "name": "occurred_from"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "Occurred From"
        - "name": "occurred_before"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "Occurred Before"
        - "name": "page_size"
          "in": "query"
          "required": false
          "schema":
            "type": "integer"
            "maximum": 200
            "minimum": 1
            "default": 50
            "title": "Page Size"
        - "name": "after_occurred_at"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
              - "type": "null"
            "title": "After Occurred At"
        - "name": "after_entry_id"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              - "type": "string"
                "format": "uuid"
              - "type": "null"
            "title": "After Entry Id"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/AuditEntryPage"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
...
"components":
  "schemas":
...
    "ActorOut":
      "properties":
        "kind":
          "type": "string"
          "title": "Kind"
        "id":
          "type": "string"
          "format": "uuid"
          "title": "Id"
        "display_name_snapshot":
          "type": "string"
          "title": "Display Name Snapshot"
        "role_in_use":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Role In Use"
        "concession_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Concession Id"
        "unit":
          "type": "string"
          "title": "Unit"
      "additionalProperties": false
      "type": "object"
      "required":
        - "kind"
        - "id"
        - "display_name_snapshot"
        - "role_in_use"
        - "concession_id"
        - "unit"
      "title": "ActorOut"
    "ApiErrorBody":
      "additionalProperties": false
      "description": "Cuerpo JSON de toda respuesta de error (domain-entities §4.4, pendiente nº 33)."
      "properties":
        "code":
          "$ref": "#/components/schemas/ApiErrorCode"
        "detail_code":
          "anyOf":
            - "maxLength": 64
              "pattern": "[a-z][a-z0-9]*(?:_[a-z0-9]+)+"
              "type": "string"
            - "type": "null"
          "default": null
          "title": "Detail Code"
        "message_es":
          "maxLength": 256
          "minLength": 1
          "title": "Message Es"
          "type": "string"
        "correlation_id":
          "format": "uuid"
          "title": "Correlation Id"
          "type": "string"
        "retry_after_seconds":
          "anyOf":
            - "maximum": 3600
              "minimum": 1
              "type": "integer"
            - "type": "null"
          "default": null
          "title": "Retry After Seconds"
      "required":
        - "code"
        - "message_es"
        - "correlation_id"
      "title": "ApiErrorBody"
      "type": "object"
    "ApiErrorCode":
      "description": "``api_error_code`` (domain-entities §1 con el pendiente nº 33): lista cerrada de U-02."
      "enum":
        - "invalid_request"
        - "not_found"
        - "unauthenticated"
        - "second_factor_required"
        - "forbidden"
        - "throttled"
        - "rate_limited"
        - "payload_too_large"
        - "conflict"
        - "temporarily_unavailable"
        - "internal_error"
        - "privacy_notice_required"
        - "period_too_long"
        - "zone_without_node"
        - "storage_unavailable"
      "title": "ApiErrorCode"
      "type": "string"
...
    "AuditCursorOut":
      "properties":
        "occurred_at":
          "type": "string"
          "title": "Occurred At"
        "entry_id":
          "type": "string"
          "format": "uuid"
          "title": "Entry Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "occurred_at"
        - "entry_id"
      "title": "AuditCursorOut"
    "AuditEntryOut":
      "properties":
        "entry_id":
          "type": "string"
          "format": "uuid"
          "title": "Entry Id"
        "chain_sequence":
          "type": "integer"
          "title": "Chain Sequence"
        "actor":
          "$ref": "#/components/schemas/ActorOut"
        "operation":
          "type": "string"
          "title": "Operation"
        "scope":
          "$ref": "#/components/schemas/ScopeOut"
        "resource_kind":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Resource Kind"
        "resource_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Resource Id"
        "filters":
          "$ref": "#/components/schemas/JsonValue"
        "result_count":
          "anyOf":
            - "type": "integer"
            - "type": "null"
          "title": "Result Count"
        "outcome":
          "type": "string"
          "title": "Outcome"
        "correlation_id":
          "type": "string"
          "format": "uuid"
          "title": "Correlation Id"
        "occurred_at":
          "type": "string"
          "title": "Occurred At"
        "previous_hash":
          "type": "string"
          "title": "Previous Hash"
        "entry_hash":
          "type": "string"
          "title": "Entry Hash"
      "additionalProperties": false
      "type": "object"
      "required":
        - "entry_id"
        - "chain_sequence"
        - "actor"
        - "operation"
        - "scope"
        - "resource_kind"
        - "resource_id"
        - "filters"
        - "result_count"
        - "outcome"
        - "correlation_id"
        - "occurred_at"
        - "previous_hash"
        - "entry_hash"
      "title": "AuditEntryOut"
    "AuditEntryPage":
      "properties":
        "items":
          "items":
            "$ref": "#/components/schemas/AuditEntryOut"
          "type": "array"
          "title": "Items"
        "next_cursor":
          "anyOf":
            - "$ref": "#/components/schemas/AuditCursorOut"
            - "type": "null"
      "additionalProperties": false
      "type": "object"
      "required":
        - "items"
        - "next_cursor"
      "title": "AuditEntryPage"
...
    "EvidenceOut":
      "properties":
        "evidence_id":
          "type": "string"
          "format": "uuid"
          "title": "Evidence Id"
        "clip_id":
          "type": "string"
          "format": "uuid"
          "title": "Clip Id"
        "camera_id":
          "type": "string"
          "format": "uuid"
          "title": "Camera Id"
        "sha256":
          "type": "string"
          "title": "Sha256"
        "size_bytes":
          "type": "integer"
          "title": "Size Bytes"
        "content_type":
          "type": "string"
          "title": "Content Type"
        "media_kind":
          "type": "string"
          "title": "Media Kind"
        "duration_ms":
          "anyOf":
            - "type": "integer"
            - "type": "null"
          "title": "Duration Ms"
        "segment":
          "type": "string"
          "title": "Segment"
        "starts_at":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Starts At"
        "ends_at":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Ends At"
        "verified_at":
          "type": "string"
          "title": "Verified At"
      "additionalProperties": false
      "type": "object"
      "required":
        - "evidence_id"
        - "clip_id"
        - "camera_id"
        - "sha256"
        - "size_bytes"
        - "content_type"
        - "media_kind"
        - "duration_ms"
        - "segment"
        - "starts_at"
        - "ends_at"
        - "verified_at"
      "title": "EvidenceOut"
      "description": "Metadatos verificados de una evidencia: nunca su clave de almacén, su URL ni sus bytes."
...
    "JsonValue": {}
...
    "LedgerRecordOut":
      "properties":
        "record_id":
          "type": "string"
          "format": "uuid"
          "title": "Record Id"
        "organization_id":
          "type": "string"
          "format": "uuid"
          "title": "Organization Id"
        "plant_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Plant Id"
        "chain_sequence":
          "type": "integer"
          "title": "Chain Sequence"
        "record_type":
          "type": "string"
          "title": "Record Type"
        "schema_version":
          "type": "integer"
          "title": "Schema Version"
        "actor":
          "$ref": "#/components/schemas/ActorOut"
        "scope":
          "$ref": "#/components/schemas/ScopeOut"
        "correlation_id":
          "type": "string"
          "format": "uuid"
          "title": "Correlation Id"
        "received_at":
          "type": "string"
          "title": "Received At"
        "occurred_at":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Occurred At"
        "source_key":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Source Key"
        "content":
          "$ref": "#/components/schemas/JsonValue"
        "content_hash":
          "type": "string"
          "title": "Content Hash"
        "previous_hash":
          "type": "string"
          "title": "Previous Hash"
        "record_hash":
          "type": "string"
          "title": "Record Hash"
        "evidences":
          "items":
            "$ref": "#/components/schemas/EvidenceOut"
          "type": "array"
          "title": "Evidences"
      "additionalProperties": false
      "type": "object"
      "required":
        - "record_id"
        - "organization_id"
        - "plant_id"
        - "chain_sequence"
        - "record_type"
        - "schema_version"
        - "actor"
        - "scope"
        - "correlation_id"
        - "received_at"
        - "occurred_at"
        - "source_key"
        - "content"
        - "content_hash"
        - "previous_hash"
        - "record_hash"
        - "evidences"
      "title": "LedgerRecordOut"
    "LedgerRecordPage":
      "properties":
        "items":
          "items":
            "$ref": "#/components/schemas/LedgerRecordOut"
          "type": "array"
          "title": "Items"
        "next_cursor":
          "anyOf":
            - "$ref": "#/components/schemas/RecordCursorOut"
            - "type": "null"
      "additionalProperties": false
      "type": "object"
      "required":
        - "items"
        - "next_cursor"
      "title": "LedgerRecordPage"
...
    "RecordCursorOut":
      "properties":
        "received_at":
          "type": "string"
          "title": "Received At"
        "record_id":
          "type": "string"
          "format": "uuid"
          "title": "Record Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "received_at"
        - "record_id"
      "title": "RecordCursorOut"
...
    "ScopeOut":
      "properties":
        "plant_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Plant Id"
        "zone_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Zone Id"
        "node_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Node Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "plant_id"
        - "zone_id"
      "title": "ScopeOut"
...